Wrap KtLint classes
### What's done:
* added DiktatRuleSet as replacement for KtLint's RuleSet
* removed extending KtLint's Rule in DiktatRule
* refactored OrderedRuleSet to KtLintRuleSetWrapper
* fixed RunInScript
