+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/rules/DiktatRule.kt
@@ -2,7 +2,6 @@ package org.cqfn.diktat.ruleset.rules
 
 import org.cqfn.diktat.common.config.rules.RulesConfig
 import org.cqfn.diktat.common.config.rules.isRuleEnabled
-import org.cqfn.diktat.common.config.rules.qualifiedWithRuleSetId
 import org.cqfn.diktat.common.utils.loggerWithKtlintConfig
 import org.cqfn.diktat.ruleset.constants.EmitType
 import org.cqfn.diktat.ruleset.utils.getFilePath
@@ -16,17 +15,16 @@ private typealias DiktatConfigRule = org.cqfn.diktat.common.config.rules.Rule
 /**
  * This is a wrapper around _KtLint_ `Rule`.
  *
- * @param id id of the rule
+ * @property id id of the rule
  * @property configRules all rules from configuration
  * @property inspections warnings that are used in the rule's code
  */
 @Suppress("TooGenericExceptionCaught")
 abstract class DiktatRule(
-    id: String,
+    val id: String,
     val configRules: List<RulesConfig>,
     private val inspections: List<DiktatConfigRule>,
-    visitorModifiers: Set<VisitorModifier> = emptySet(),
-) : Rule(id.qualifiedWithRuleSetId(), visitorModifiers) {
+) {
     /**
      * Default value is false
      */
@@ -48,7 +46,7 @@ abstract class DiktatRule(
     lateinit var emitWarn: EmitType
 
     @Suppress("TooGenericExceptionThrown")
-    final override fun visit(
+    fun visit(
         node: ASTNode,
         autoCorrect: Boolean,
         emit: EmitType
