+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/util/FixTestBase.kt
@@ -5,6 +5,7 @@
 package org.cqfn.diktat.util
 
 import org.cqfn.diktat.common.config.rules.RulesConfig
+import org.cqfn.diktat.ruleset.rules.DiktatRule
 import org.cqfn.diktat.ruleset.utils.LintErrorCallback
 import org.cqfn.diktat.ruleset.utils.defaultCallback
 import org.cqfn.diktat.ruleset.utils.format
@@ -22,7 +23,7 @@ import kotlin.io.path.div
  */
 open class FixTestBase(
     resourceFilePath: String,
-    ruleSupplier: (rulesConfigList: List<RulesConfig>) -> Rule,
+    ruleSupplier: (rulesConfigList: List<RulesConfig>) -> DiktatRule,
     defaultRulesConfigList: List<RulesConfig>? = null,
     cb: LintErrorCallback = defaultCallback,
 ) {
